+++ src/effects/Move.cpp
@@ -4,7 +4,7 @@
 #include "events/BattleField.h"
 #include "effects/Move.h"
 namespace effects {
-    Move::Move(std::weak_ptr<entity::Entity> entity, std::weak_ptr<BattleField> battleField, int offset) : ImmediateEffect(entity, std::map<size_t, int>{}, 0, 0) {
+    Move::Move(std::weak_ptr<entity::Entity> entity, std::weak_ptr<BattleField> battleField, int offset) : ImmediateEffect(entity, std::map<int, int>{}, 0, 0) {
         m_battleField = battleField;
         m_offset = offset;
     }
